+ moved the dummy values of the popup into the popup component
